move most definitions from papi header to implementation file
